+++ apps/samples/cv2x_tcp_app/Cv2xTcpApp.cpp
@@ -159,7 +159,7 @@ static void createTcpSocketCallback(shared_ptr<ICv2xTxRxSocket> sock,
 static uint64_t getCurrentTimestamp(void) {
     struct timeval tv;
     gettimeofday(&tv, NULL);
-    return tv.tv_sec * 1000000ull + tv.tv_usec;
+    return (uint64_t)tv.tv_sec * 1000000ull + (uint64_t)tv.tv_usec;
 }
 
 // Fills buffer with dummy data
